Create routes

+++ app/Http/Requests/StoreAddressRequest.php
@@ -20,12 +20,12 @@ class StoreAddressRequest extends FormRequest
     public function rules(): array
     {
         return [
-            'number' => ['required', 'integer', 'min:1', 'max:9999999'],
+            'number' => ['required', 'numeric', 'min:1', 'max:9999999'],
             'complement' => ['sometimes', 'string', 'max:255'],
             'neighborhood' => ['required', 'string', 'max:255'],
             'street' => ['required', 'string', 'max:255'],
             'zip_code' => ['required', 'string', 'size:8'],
-            'external_city_id' => ['required', 'integer'],
+            'external_city_id' => ['required', 'numeric'],
         ];
     }
 }
